split the file
