+++ greengrocer_main_class.rb
@@ -11,17 +11,7 @@
 # 個数を選択
 # 合計を計算
 
-
-class Product
-  attr_reader :id, :name, :price 
-  @@count = 30
-
-  def initialize(products_params)
-     @id = @@count += 1
-    @name = products_params[:name]
-    @price = products_params[:price]
-  end
-end
+require "./product.rb"
 
 class Greengrocer
   attr_reader :products
